refactor: split monolithic API file into focused modules
