+++ src/features/tracks/lib/metadata.ts
@@ -1,4 +1,5 @@
-import { api } from "../api/api";
+import { api } from "@/features/tracks/api";
+
 import { SearchParams } from "../types";
 
 export async function generateTrackMetadata(trackId: string) {
